refactor: move functions to different modules
