+++ commitcanvas/generate_type/classify.py
@@ -26,63 +26,3 @@
 #     else:
 #         rf.model(name,language)
 
-import sys
-from commitcanvas.generate_type.tokenizers import stem_tokenizer
-from commitcanvas.generate_type.tokenizers import dummy
-from commitcanvas.generate_type import commit_data as cm
-import joblib
-import pandas as pd
-import shutil
-from subprocess import check_output
-import io
-import pkg_resources
-
-def staged_stats(diff,commit_subject):
-
-    decoded_diff = diff.decode('utf-8')
-
-    # convert to string output into pandas dataframe for easier calculations
-    data = io.StringIO(decoded_diff)
-    df = pd.read_csv(data, sep="\t",names = ["added","deleted","file_paths"])
-
-
-    added = df.added.sum()
-    deleted = df.deleted.sum()
-    paths = df.file_paths.tolist()
-    file_extensions = cm.get_file_extensions(paths)
-    test_files_count = cm.test_files(paths)
-
-
-    staged_changes_stats = {
-        'commit_subject': commit_subject,
-        "num_files": len(paths),
-        "test_files": test_files_count,
-        "test_files_ratio": cm.get_ratio(test_files_count,paths),
-        "unique_file_extensions": file_extensions,
-        "num_unique_file_extensions": len(file_extensions),
-        "num_lines_added": added,
-        "num_lines_removed": deleted,
-        "num_lines_total": added + deleted,    
-    }
-
-    return (pd.DataFrame([staged_changes_stats]))
-
-
-
-def entry():
-    """Get commit message from command line and do checks."""
-
-    commit_msg_filepath = sys.argv[1]
-    #commitcanvas_check.commit_check(commit_msg_filepath)
-    diff = check_output(['git', 'diff', '--staged','--numstat']).strip()
-    
-    # Figure out which branch we're on
-    with open(commit_msg_filepath, 'r+') as f:
-        content = f.read()
-        # rewrite the file so that you avoid the seek
-        f.seek(0, 0)
-        stats = staged_stats(diff, content)
-        my_data = pkg_resources.resource_stream(__name__, "model/trained_model.pkl")
-        model = joblib.load(my_data)
-        predicted = model.predict(stats)[0]
-        f.write("{}: {}".format(predicted,content))
