+++ pyproject.toml
@@ -23,7 +23,7 @@ pytest-cov = "^2.10.0"
 codecov = "^2.1.8"
 
 [tool.poetry.scripts]
-commitcanvas = "commitcanvas.generate_type.classify:entry"
+commitcanvas = "commitcanvas.__main__:entry"
 #save = "commitcanvas.generate_type.data_processing:save"
 
 # [tool.poetry.plugins."blogtool.parsers"]
